move dedup paths to its own file
